acl: API cleanup
Use consistent API types.

Type: fix

+++ src/vnet/ip/ip_types_api.h
@@ -30,9 +30,9 @@
  * These enum decode/encodes use 'int' as the type for the enum because
  * one cannot forward declare an enum
  */
-extern int ip_address_family_decode (int _af, ip_address_family_t * out);
+extern int ip_address_family_decode (u8 _af, ip_address_family_t * out);
 extern int ip_address_family_encode (ip_address_family_t af);
-extern int ip_proto_decode (int _af, ip_protocol_t * out);
+extern int ip_proto_decode (u8 _af, ip_protocol_t * out);
 extern int ip_proto_encode (ip_protocol_t af);
 extern ip_dscp_t ip_dscp_decode (u8 _dscp);
 extern u8 ip_dscp_encode (ip_dscp_t dscp);
